[green] split tests
